Yet another attempt at dealing with prompts. The most succesful one so far....
Yet another attempt at dealing with prompts.

(*) We suppress all FICS generated prompts.
(*) After a board we always send a prompt in the same write.
xboard likes it that way.
(*) After feedback on our own commands we always send a prompt.
(*) After unsollicited FICS input we look in the input
buffer. If we see a prompt then we append one to our write.