A new attempt at fixing the newline issue on Windows.
The newline issue on Windows is caused by the fact
that timeseal.exe generates \r\n\r as eol sequence.
This uncommon eol sequence confused icsdrone.

With the new algorithm every sequence of consecutive \r and \n
characters, containing at least one \n is considered to be
an eol sequence.

This should also work when not using timeseal.exe and on
Linux where the eol sequence is \n\r (which is also very
strange BTW).