Make it possible to kibitz against computers but not humans
Done indirectly by adding -sendComputerGame and -sendHumanGame
flags. With these icsdrone can send instructions to the server at
the start of each game, specific for the type of opponent.  This
can then set or unset aliases (or be used for other purposes).

Note that just varying the "feedbackCommand" wouldn't work, because
that doesn't catch the chats coming out of Polyglot.